renamed _test files to _verifiers
