+++ verifiers/optimization_verifier.py
