+++ verifiers/verifier_runner.py
@@ -3,17 +3,13 @@ import os
 import pickle
 import random
 from os.path import join
-import sys
 from pathlib import Path
-sys.path.append('../cpmpy')
-import cpmpy as cp
-from cpmpy.exceptions import CPMpyException
 from mutators import *
-from .metamorphic_test import Metamorphic_Verifier
-from .solution_test import Solution_Verifier
-from .model_counting_test import Model_Count_Verifier
-from .equivalance_test import Equivalance_Verifier
-from .optimization_test import Optimization_Verifier
+from .metamorphic_verifier import Metamorphic_Verifier
+from .solution_verifier import Solution_Verifier
+from .model_counting_verifier import Model_Count_Verifier
+from .equivalance_verifier import Equivalance_Verifier
+from .optimization_verifier import Optimization_Verifier
 import math
 import warnings
 import traceback
@@ -29,6 +25,9 @@ def create_error_output_text(error_data: dict) -> str:
             error_data (dict): the dict containing all the info about the error that occured
     """
     execution_time_text = str(math.floor(error_data["execution_time"]/60)) + " minutes " + str(math.floor(error_data["execution_time"]%60)) + " seconds"
+    verifier_text = ""
+    if error_data["error"]["type"] != "fuzz_test_crash":
+        verifier_text = "Chosen Verifier: "+error_data["verifier"]
     error_text = ""
     # get all the error details
     for key, value in error_data["error"].items():
@@ -44,6 +43,7 @@ def create_error_output_text(error_data: dict) -> str:
 An error occured while running a test
 
 Used solver: {solver}
+{verifier_text}
 With {mutations} mutations per model
 With seed: {seed}
 The test failed in {execution_time}
@@ -51,7 +51,7 @@ The test failed in {execution_time}
 Error Details:
 {error_text}
     \
-    '''.format(solver=error_data["solver"], mutations=error_data["mutations_per_model"],seed =error_data["seed"],execution_time=execution_time_text,error_text=error_text)
+    '''.format(verifier_text=verifier_text,solver=error_data["solver"], mutations=error_data["mutations_per_model"],seed =error_data["seed"],execution_time=execution_time_text,error_text=error_text)
     
 
 
@@ -113,7 +113,7 @@ def run_verifiers(current_amount_of_tests, current_amount_of_error, lock, solver
             if not (error == None) and current_amount_of_error.value < max_error_treshold:
                 lock.acquire()
                 try:
-                    error_data = {'solver' : solver, 'mutations_per_model' : mutations_per_model, "seed": random_seed, "execution_time": execution_time, "error" :error}
+                    error_data = {'verifier':random_verifier.getName(),'solver' : solver, 'mutations_per_model' : mutations_per_model, "seed": random_seed, "execution_time": execution_time, "error" :error}
                     write_error(error_data,output_dir)
                     current_amount_of_error.value +=1
                 finally:
